requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
wikipedia>=1.4.0 